// src/lib/sanity/sanity.types.ts
export type ExperienceEntriesQueryResult = Array<{
  _id: string;
  _type: "experienceEntry";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  dateRange: string;
  location?: string;
  role: string;
  description: string;
  dotColor: "magenta" | "primary" | "secondary" | "soft" | "tertiary";
  order: number;
  compact?: boolean;
}>;